+++ Source/NSSocketPort.m
@@ -2030,7 +2030,7 @@ static Class		tcpPortClass;
   [d handlePortMessage: m];
 }
 
-- (unsigned) hash
+- (NSUInteger) hash
 {
   return (unsigned)portNum;
 }
